launch_trainer.py further refactoring
